Move anchor and canvas helpers into core
